+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## [0.8.0rc4](https://github.com/dldevinc/paper-uploads/tree/v0.8.0rc4) - 2021-02-08
+### Features
+- Management commands rewritten.
+- Added `create_missing_variations` management command.
+
 ## [0.8.0rc3](https://github.com/dldevinc/paper-uploads/tree/v0.8.0rc3) - 2021-02-02
 ### ⚠ BREAKING CHANGES
 - Method `set_owner_from()` has been rewritten and renamed to `set_owner_field()`.
@@ -8,6 +13,8 @@
 - `SizeValidator` is now deprecated in favor of `MaxSizeValidator`.
 - Allow `str` and `Path` as the parameter of `attach()` method.
 - Allow overriding variation name with option `name`.
+### Bug Fixes
+- Fixed `remove_variations` and `recreate_variations` management commands.
 
 ## [0.8.0rc2](https://github.com/dldevinc/paper-uploads/tree/v0.8.0rc2) - 2021-02-01
 ### Features